[fixes #18652909] Replace flash charts with google ones; remove unused js
