+++ app/controllers/application_controller.rb
@@ -32,12 +32,6 @@ class ApplicationController < ActionController::Base
       end
     end
 
-    # load vars for pretty charts
-    def load_charts
-      @response = @data_response    = current_response
-      @projects                     = @data_response.projects.find(:all, :order => "name ASC") if @data_response
-    end
-
   private
 
     def invalid_method
