+++ app/controllers/reporter/reports_controller.rb
@@ -1,15 +1,9 @@
 class Reporter::ReportsController < Reporter::BaseController
   include ReportsControllerHelpers
-
-  before_filter :load_charts, :only => :index
+  include PrepareCharts
 
   def index
-    @activities_without_projects  = @data_response.activities.roots.without_a_project
-    @other_costs_without_projects = @data_response.other_costs.without_a_project
-    @code_roots                   = Code.purposes.roots
-    @cost_cat_roots               = CostCategory.roots
-    @other_cost_roots             = OtherCostCode.roots
-    @policy_maker                 = true #view helper
+    load_dashboard_charts
   end
 
   def show
