+++ app/stylesheets/partials/_lists.sass
@@ -196,7 +196,7 @@ ul.review li
 ul.group
   border-bottom: 1px solid $light-blue
   margin-top: 10px
-  margin-bottom: 20px
+  margin-bottom: 20px !important
   font-size: 11px
   +linear-gradient(color-stops(white, $lightest-blue + #090909))
