--- app/views/shared/data_responses/_chart_nav.html.haml
@@ -1,19 +1,0 @@
-.tabs_nav.full
-
-  %ul.inline_tab
-  
-    %li{:id => 'tab1', :class => 'selected'}
-      = link_to "MTEF", "#"
-    
-    %li{:id => 'tab2'}
-      = link_to "NSP", "#"
-
-    %li{:id => 'tab3'}
-      = link_to "Inputs", "#"
-
-    %li{:id => 'tab4'}
-      = link_to "HSSPII Strategic Program", "#"
-
-    %li{:id => 'tab5'}
-      = link_to "HSSPII Strategic Objective", "#"
-
