--- app/views/shared/data_responses/_chart_tab.html.haml
@@ -1,6 +1,0 @@
-/ = render 'shared/data_responses/codings_inline_tab', :chart_type => chart_type, :link_name => 'Pie'
-
-.pie
-  .chart{:id => "#{element_type}_#{element.id}_#{chart_type}_pie"}
-.tree
-  .chart{:id => "#{element_type}_#{element.id}_#{chart_type}_tree", :style => "width:600px; height:400px;"}
