+++ app/views/shared/data_responses/_data_responses_charts.html.haml
@@ -1,34 +1,15 @@
-= render '/shared/data_responses/chart_nav'
+%ul.group
+  %li= link_to "NSP", {:code_type => "nsp"}, active_if(@nsp)
+  %li= link_to "MTEF", {:code_type => "mtef"}, active_if(@mtef)
+  %li= link_to "Inputs", {:code_type => "cost_category"}, active_if(@cost_category)
+  %li= link_to "HSSP2 Strat Prog", {:code_type => "hssp2_strat_prog"}, active_if(@hssp2_strat_prog)
+  %li= link_to "HSSP2 Strat Obj", {:code_type => "hssp2_strat_obj"}, active_if(@hssp2_strat_obj)
 
+.full
+  .first-half
+    %h2= "#{@chart_name} Expenditure"
+    #code_spent.bordered
 
-.tabs.full
-  .tab1{:"data-chart_type" => "data_response", :"data-id" => data_response.id, :style => "float:left"}
-    = render 'shared/data_responses/chart_tab', :chart_type => 'mtef_spend', :element => data_response, :element_type => 'data_response'
-    
-  .tab1{:"data-chart_type" => "data_response", :"data-id" => data_response.id, :style => "float:right"}
-    = render 'shared/data_responses/chart_tab', :chart_type => 'mtef_budget', :element => data_response, :element_type => 'data_response'
-
-  .tab2{:"data-chart_type" => "data_response", :"data-id" => data_response.id, :style => "float:left"}
-    = render 'shared/data_responses/chart_tab', :chart_type => 'nsp_spend', :element => data_response, :element_type => 'data_response'
-
-  .tab2{:"data-chart_type" => "data_response", :"data-id" => data_response.id, :style => "float:right"}
-    = render 'shared/data_responses/chart_tab', :chart_type => 'nsp_budget', :element => data_response, :element_type => 'data_response'
-
-  .tab3{:"data-chart_type" => "data_response", :"data-id" => data_response.id, :style => "float:left"}
-    = render 'shared/data_responses/chart_tab', :chart_type => 'cc_spend', :element => data_response, :element_type => 'data_response'
-
-  .tab3{:"data-chart_type" => "data_response", :"data-id" => data_response.id, :style => "float:right"}
-    = render 'shared/data_responses/chart_tab', :chart_type => 'cc_budget', :element => data_response, :element_type => 'data_response'
-    
-  .tab4{:"data-chart_type" => "data_response", :"data-id" => data_response.id, :style => "float:left"}
-    = render 'shared/data_responses/chart_tab', :chart_type => 'stratprog_spend', :element => data_response, :element_type => 'data_response'
-
-  .tab4{:"data-chart_type" => "data_response", :"data-id" => data_response.id, :style => "float:right"}
-    = render 'shared/data_responses/chart_tab', :chart_type => 'stratprog_budget', :element => data_response, :element_type => 'data_response'
-    
-  .tab5{:"data-chart_type" => "data_response", :"data-id" => data_response.id, :style => "float:left"}
-    = render 'shared/data_responses/chart_tab', :chart_type => 'stratobj_spend', :element => data_response, :element_type => 'data_response'
-
-  .tab5{:"data-chart_type" => "data_response", :"data-id" => data_response.id, :style => "float:right"}
-    = render 'shared/data_responses/chart_tab', :chart_type => 'stratobj_budget', :element => data_response, :element_type => 'data_response'
-    
+  .second-half
+    %h2= "#{@chart_name} Current Budget"
+    #code_budget.bordered
