--- app/views/shared/data_responses/_javascript.html.haml
@@ -1,7 +1,0 @@
-- content_for :head do
-  <script type="text/javascript" src="https://www.google.com/jsapi"></script>
-
-:javascript
-  _projects = #{projects.map(&:id).to_json};
-  _dr_id = "#{data_response.id}";
-  google.load("visualization", "1", {packages:["treemap"]});
