+++ app/views/shared/data_responses/_show.html.haml
@@ -1,5 +1,11 @@
-- javascript "/ampie/swfobject.js"
+- content_for :head do
+  <script type="text/javascript" src="https://www.google.com/jsapi"></script>
 
-= render 'shared/data_responses/javascript', :projects => @projects, :data_response => @response
+:javascript
+  google.load("visualization", "1", {packages:["corechart"]});
+  var _spent_pie_values   = #{js_safe(@spent_pie_values)};
+  var _budget_pie_values  = #{js_safe(@budget_pie_values)};
+  var _code_spent_values  = #{js_safe(@code_spent_values)};
+  var _code_budget_values = #{js_safe(@code_budget_values)};
 
 = render 'shared/data_responses/data_responses_charts', :data_response => @response
